Add 2020/05
The initial solution is 2020/05/solve.py.

The concept of taking a range, cutting it in half and then taking
either the top or bottom half according to a binary choice is
precisely how a binary number is represented. For example, taking the
range 0-7, then taking right (1), left (0), then right (1) would
produce 5, which is the same as 0b101. This is because takig the
"right" of 0-7 gives 4-7, then "left" gives 4-5, then "right" gives 5.