Add solutions for 2019 (up to day 19)
Note for day 16: -

This day was not solved completely without help.  Step 1 was solved by pure
Python however it was slow.  Step 2 however required help, specifically to
realise that the pattern for the repeated signal will be only 1s and 0s,
meaning that the calculation can be vastly simplified.

When searching for other solutions, the use of Numpy was investigated.  This
eventually resulted in the cumsum() trick and optimisation of part 1.
Extensive notes have been added to solve.py to document the learning process.